requirements.txt
allure-combine==1.0.11
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
enum34==1.1.10
exceptiongroup==1.2.2
idna==3.7
iniconfig==2.0.0
lxml==5.2.2
namedlist==1.8
packaging==24.1
pluggy==1.5.0
pytest==8.3.1
pytest-rerunfailures==14.0
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
soupsieve==2.5
tomli==2.0.1
urllib3==2.2.2

websockets~=12.0
pyotp~=2.9.0